selialization, write and read files, sorts, time manage done

+++ src/exceptions/InvalidSelectionException.java
@@ -1,14 +1,14 @@
 package exceptions;
 
 public class InvalidSelectionException extends Exception {
-	private int selection;
+	private double selection;
 	
-	public InvalidSelectionException(int selection) {
+	public InvalidSelectionException(double selection) {
 		super("Your selection: "+selection+". Is not valid");
 		this.selection = selection;
 	}
 	
-	public int getSelection() {
+	public double getSelection() {
 		return selection;
 	}
 }
